@@ -10,23 +10,23 @@
   },
   "description": "Nuxt Cookie Control Module",
   "devDependencies": {
-    "@commitlint/cli": "18.4.4",
-    "@commitlint/config-conventional": "18.4.4",
+    "@commitlint/cli": "19.2.1",
+    "@commitlint/config-conventional": "19.1.0",
     "@dargmuesli/nuxt-cookie-control": "link:",
     "@nuxt/module-builder": "0.5.5",
-    "@nuxt/schema": "3.9.0",
+    "@nuxt/schema": "3.11.1",
     "@nuxtjs/eslint-config-typescript": "12.1.0",
-    "eslint": "8.56.0",
+    "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-prettier": "5.1.2",
-    "husky": "8.0.3",
-    "lint-staged": "15.2.0",
-    "nuxt": "3.9.0",
-    "prettier": "3.1.1",
-    "typescript": "5.3.3",
-    "vue": "3.4.5",
-    "vue-tsc": "1.8.27",
-    "webpack": "5.89.0"
+    "eslint-plugin-prettier": "5.1.3",
+    "husky": "9.0.11",
+    "lint-staged": "15.2.2",
+    "nuxt": "3.11.1",
+    "prettier": "3.2.5",
+    "typescript": "5.4.2",
+    "vue": "3.4.21",
+    "vue-tsc": "2.0.6",
+    "webpack": "5.90.3"
   },
   "engines": {
     "node": ">=16"
@@ -61,20 +61,20 @@
     "Jonas Thelemann"
   ],
   "name": "@dargmuesli/nuxt-cookie-control",
-  "packageManager": "pnpm@8.14.0",
+  "packageManager": "pnpm@8.15.5",
   "publishConfig": {
     "access": "public"
   },
   "repository": "https://github.com/dargmuesli/nuxt-cookie-control",
   "resolutions": {
-    "@nuxt/kit": "3.9.0"
+    "@nuxt/kit": "3.11.1"
   },
   "scripts": {
     "build": "nuxt-module-build build",
     "lint": "eslint --ext .js,.ts,.vue . && nuxt typecheck",
     "lint:fix": "eslint --ext .js,.ts,.vue --fix . && nuxt typecheck --fix",
     "prepack": "pnpm build",
-    "prepare": "husky install"
+    "prepare": "husky"
   },
   "type": "module",
   "types": "./dist/module.d.ts",
