@@ -1,11 +1,11 @@
-const commitlintConfigConventional = require("@commitlint/config-conventional");
+import commitlintConfigConventional from "@commitlint/config-conventional"
 
 const ruleMaxLineLength =
   commitlintConfigConventional.rules["body-max-line-length"];
 
 ruleMaxLineLength[0] = process.env.CI === "true" ? 1 : 2;
 
-module.exports = {
+export default {
   extends: ["@commitlint/config-conventional"],
   rules: {
     "body-max-line-length": ruleMaxLineLength,
